+++ lib/Alembic/AbcGeom/Tests/HelperLibTest.cpp
@@ -58,9 +58,9 @@ using Alembic::Util::uint32_t;
 
 void writeSimpleProperties(const std::string &archiveName)
 {
-    const unsigned int numChildren = 3;
+    const int numChildren = 3;
 
-    const unsigned int numSamples = 5;
+    const int numSamples = 5;
     const chrono_t dt = 1.0 / 24.0;
 
     TimeSampling ts(dt, 666.0); // uniform with cycle=dt starting at 666.0
@@ -127,7 +127,7 @@ void readSimpleProperties(const std::string &archiveName)
     IObject archiveTop = archive.getTop();
 
     // Determine the number of (top level) children the archive has
-    const unsigned int numChildren = archiveTop.getNumChildren();
+    const int numChildren = archiveTop.getNumChildren();
     TESTING_ASSERT( numChildren == 4 );
     std::cout << "The archive has " << numChildren << " children:"
               << std::endl;
@@ -149,7 +149,7 @@ void readSimpleProperties(const std::string &archiveName)
 
         // Properties
         ICompoundProperty props = child.getProperties();
-        size_t numProperties = props.getNumProperties();
+        int numProperties = props.getNumProperties();
 
         std::cout << "  ..and " << numProperties << " simple properties"
                   << std::endl;
@@ -242,7 +242,7 @@ void readSimpleProperties(const std::string &archiveName)
                 GetCompoundPropertyReaderPtr(props)->
                 getScalarProperty( propNames[jj] )->getTimeSampling();
 
-            size_t numSamples = ts->getNumStoredTimes();
+            int numSamples = ts->getNumStoredTimes();
 
 
             std::cout << "    ..and "
@@ -373,7 +373,7 @@ void readSimpleProperties(const std::string &archiveName)
 
 void writeNestedCommpoundWithVis(const std::string &archiveName)
 {
-    const unsigned int numChildren = 2;
+    const int numChildren = 2;
 
     // Create an archive for writing. Indicate that we want Alembic to
     //   throw exceptions on errors.
@@ -429,7 +429,6 @@ void readNestedCommpoundWithVis(const std::string &archiveName)
                       archiveName, ErrorHandler::kThrowPolicy );
     IObject archiveTop = archive.getTop();
 
-
     ICharProperty topVisibility = GetVisibilityProperty (archiveTop);
     std::cout << "Does this object have VisibilityProperty? " 
               << (bool) (true == topVisibility)
@@ -437,7 +436,7 @@ void readNestedCommpoundWithVis(const std::string &archiveName)
     ABCA_ASSERT( topVisibility == false, "top object should not have a visibility property");
 
     // Determine the number of (top level) children the archive has
-    const unsigned int numChildren = archiveTop.getNumChildren();
+    const int numChildren = archiveTop.getNumChildren();
     ABCA_ASSERT( numChildren == 2, "Wrong number of children (expected 2)");
     std::cout << "The archive has " << numChildren << " children:"
               << std::endl;
@@ -455,7 +454,7 @@ void readNestedCommpoundWithVis(const std::string &archiveName)
 
         // Properties
         ICompoundProperty props = child.getProperties();
-        size_t numProperties = props.getNumProperties();
+        int numProperties = props.getNumProperties();
 
         std::cout << "  ..and " << numProperties << " properties"
                   << std::endl;
